refactored to put date options on main app and pass it down to each component
