+++ lib/components/Admin.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import firebase from '../firebase';
-import { map, uniq } from 'lodash';
+import { map } from 'lodash';
 import AdminSpike from './AdminSpike';
 import Spike from './Spike';
 
@@ -29,24 +29,13 @@ export default class Admin extends Component {
     })
   }
 
-  dateOptions(spikes) {
-    let dates = uniq(map(spikes, (spike) => spike.spikeDate)).sort((a,b) => {
-      a = a.split('-').join('');
-      b = b.split('-').join('');
-      return a > b ? 1 : a < b ? -1 : 0;
-    })
-    return map(dates, (date, index) => {
-      return <option key={index} value={date}>{date}</option>
-    })
-  }
-
   render() {
     const { spikes, updateSpike, deleteSpike, admins,
-      newAdmin, removeAdmin } = this.props;
+      newAdmin, removeAdmin, dateOptions } = this.props;
 
     let allSpikes = this.filterSpikes(spikes)
 
-    let allDateOptions = this.dateOptions(spikes)
+    let allDateOptions = dateOptions(spikes)
 
     let allAdmins = map(admins, (admin, index) => {
       return (
