+++ lib/components/App.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { map, extend, invert } from 'lodash';
+import { map, extend, invert, uniq } from 'lodash';
 import firebase, { signIn, signOut, reference, adminReference } from '../firebase';
 import SignIn from './SignIn';
 import Spikes from './Spikes';
@@ -106,6 +106,17 @@ export default class App extends Component {
     firebase.database().ref(`spikes/${spike.key}`).update( spike );
   }
 
+  dateOptions(spikes) {
+    let dates = uniq(map(spikes, (spike) => spike.spikeDate)).sort((a,b) => {
+      a = a.split('-').join('');
+      b = b.split('-').join('');
+      return a > b ? 1 : a < b ? -1 : 0;
+    })
+    return map(dates, (date, index) => {
+      return <option key={index} value={date}>{date}</option>
+    })
+  }
+
   render() {
     const user = this.state.user;
     if (!user) { return <SignIn /> }
@@ -127,6 +138,7 @@ export default class App extends Component {
             deleteSpike={(spike) => this.deleteSpike(spike)}
             newAdmin={(email) => this.newAdmin(email)}
             removeAdmin={(index) => this.removeAdmin(index)}
+            dateOptions={this.dateOptions}
           />
         </section>
       )
@@ -146,6 +158,7 @@ export default class App extends Component {
             spikes={this.state.spikes}
             user={user}
             updateCount={(spike) => this.updateCount(spike)}
+            dateOptions={this.dateOptions}
           />
         </section>
       )
