+++ lib/components/Spike.js
@@ -56,18 +56,17 @@ const Spike = ({ spike, createSpike, updateCount, toggleForm, user }, key, admin
             />
             </label>
             <section className="ButtonContainer">
-              <input
+              <button
                 className='SubmitButton'
                 type='submit'
-              />
-              <input
-                value='Cancel'
+              >Submit</button>
+              <button
                 className='CancelButton'
                 onClick={(e) => {
                   e.preventDefault();
                   toggleForm(e)
                 }}
-              />
+              >Cancel</button>
             </section>
           </form>
         </section>
