+++ lib/components/Spikes.js
@@ -1,6 +1,6 @@
 import React , { Component } from 'react';
 import firebase from '../firebase';
-import { map, uniq } from 'lodash';
+import { map } from 'lodash';
 import Spike from './Spike';
 
 
@@ -31,23 +31,12 @@ export default class Spikes extends Component {
     })
   }
 
-  dateOptions(spikes) {
-    let dates = uniq(map(spikes, (spike) => spike.spikeDate)).sort((a,b) => {
-      a = a.split('-').join('');
-      b = b.split('-').join('');
-      return a > b ? 1 : a < b ? -1 : 0;
-    })
-    return map(dates, (date, index) => {
-      return <option key={index} value={date}>{date}</option>
-    })
-  }
-
   render() {
-    const { spikes } = this.props
+    const { spikes, dateOptions } = this.props
 
     let allSpikes = this.filterSpikes(spikes)
 
-    let allDateOptions = this.dateOptions(spikes)
+    let allDateOptions = dateOptions(spikes)
 
     return (
       <section>
